// src/app/auth/model/interco-dto.ts
export class IntercoDto {

    prenom?: string;
    nom?: string;
    email?: string;
    password?: String;
    nin?: string;
    telephone?: string;
    datedenaissance?: string;
    lieudenaissance?: string;
    adresse?: string;
    sexe?: string;
   fonction?: string;
   scannernin?: string;
   confirmemail?: string;
}